move cpuid stuff to util/arch/x86
