+++ CMakeLists.txt
@@ -564,8 +564,8 @@ install(FILES ${hs_HEADERS} DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/hs")
 set (hs_exec_common_SRCS
     src/alloc.c
     src/scratch.c
-    src/util/cpuid_flags.c
-    src/util/cpuid_flags.h
+    src/util/arch/x86/cpuid_flags.c
+    src/util/arch/x86/cpuid_flags.h
     src/util/multibit.c
     )
