+++ src/util/target_info.cpp
@@ -29,7 +29,9 @@
 
 #include "hs_compile.h" // for various hs_platform_info flags
 #include "target_info.h"
-#include "util/cpuid_flags.h"
+#if defined(ARCH_X86_64)
+#include "util/arch/x86/cpuid_flags.h"
+#endif
 
 namespace ue2 {
